Just added NINT() instead of implicit conversion after first LAPACK call that returns desired array size in a real variable.
	modified:   multigrid.f90

+++ multigrid.f90
@@ -790,7 +790,8 @@ integer,save:: nx,ny,nz,nulx,nuly,nulz,nxyz,called=0
      write (*,*) "info",info
     endif
 
-    ldwork=work(1)
+    ldwork=nint(work(1))
+
     deallocate(work)
     allocate(work(ldwork))
